fixup! Add Snowflake JDBC Connector
