+++ testing/trino-product-tests-launcher/src/main/resources/docker/presto-product-tests/conf/environment/multinode-all/snowflake.properties
@@ -1,4 +1,4 @@
 connector.name=snowflake
-connection-url=${ENV:SNOWFLAKE_URL}
-connection-user=${ENV:SNOWFLAKE_USER}
-connection-password=${ENV:SNOWFLAKE_PASSWORD}
+connection-url=jdbc:snowflake://example.snowflakecomputing.com
+connection-user=root
+connection-password=secret
