+++ testing/trino-product-tests-launcher/src/main/resources/docker/presto-product-tests/conf/environment/multinode-snowflake/snowflake.properties
@@ -2,3 +2,6 @@ connector.name=snowflake
 connection-url=${ENV:SNOWFLAKE_URL}
 connection-user=${ENV:SNOWFLAKE_USER}
 connection-password=${ENV:SNOWFLAKE_PASSWORD}
+snowflake.database=${ENV:SNOWFLAKE_DATABASE}
+snowflake.role=${ENV:SNOWFLAKE_ROLE}
+snowflake.warehouse=${ENV:SNOWFLAKE_WAREHOUSE}
